Properly create and use the OVS plugin.
Instead leaving the prepackaged OVS plugin unchanged, actually build
it up according to how we want to configure Quantum.

This also gets rid of the dirty hack we were using of symlinking the
ovs .ini to quantum.conf.

While we are at it, make sure we default to flat mode instead none.